+++ apx_sdk/CHANGELOG.md
@@ -16,6 +16,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Moved `Url` type from `apx_sdk` to `apx_core` package.
 
+### Removed
+
+- Removed `FromStr` implementation from `Url` type.
+
 ## [0.12.0] - 2025-04-24
 
 ### Added